\docType{class}
\name{cap-class}
\alias{cap-class}
\title{Class \code{"cap"} in Package \pkg{HandTill2001}}
\description{
  A virtual class for \code{"bincap"} and \code{"multcap"}.
}
\section{Objects from the Class}{
  A virtual Class: No objects may be created from it.
}
\author{
  Andreas Dominik Cullmann
}
\seealso{
  \code{"\link[=bincap-class]{class?bincap}"},
  \code{"\link[=multcap-class]{class?multcap}"}
}
\keyword{classes}

