\docType{class}
\name{multcap-class}
\alias{multcap-class}
\title{Class \code{"multcap"} in Package \pkg{HandTill2001}}
\description{
  S4 class for a multiple class response and corresponding
  (predicted) probabilities.
}
\section{Objects from the Class}{
  Objects can be created by calls of the form
  \code{new("multcap", ...)}. They are used to store a
  multiple class response and the predicted probabilities
  for each of the \code{levels(response)}.
}
\examples{
showClass("multcap")
}
\author{
  Andreas Dominik Cullmann
}
\seealso{
  \code{"\link[=cap-class]{class?cap}"} ,
  \code{"\link[=bincap-class]{class?bincap}"} ,
  \code{"\link[=multcap]{?multcap}"}
}
\keyword{classes}

