% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{cap-class}
\alias{cap-class}
\title{Class and Prediction Objects}
\description{
A virtual class for \code{bincap} and \code{multcap}.
}
\section{Objects from the Class}{
 A virtual Class: No objects may be created
from it.
}

\seealso{
\code{\link[=bincap-class]{class?HandTill2001::bincap}},
\code{\link[=multcap-class]{class?HandTill2001::multcap}}
}
\keyword{classes}
\keyword{internal}
