% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{multcap-class}
\alias{multcap-class}
\title{Multiple Class and Prediction Objects}
\description{
S4 class for a multiple class response and corresponding (predicted)
probabilities.
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("multcap", ...)}. They are used to store a multiple class response
and the predicted probabilities for each of the \code{levels(response)}.
}

\examples{

showClass("multcap")
}
\seealso{
\code{\link[=cap-class]{class?HandTill2001::cap}} ,
\code{\link[=bincap-class]{class?HandTill2001::bincap}} ,
\code{\link[=multcap]{?HandTill2001::multcap}}
}
\keyword{classes}
